Add unit test for --group-failures option.
In doing so, we needed to sort the tests before collecting their failure
messages.  This ensures a consistent ordering of the failure messages,
making it easier to verify that the expected output is contained within
the group markers.

Since the test data only consists of a single file, this doesn't verify
that failures are grouped by year.  However, it's better than nothing.